Resource protection should be deep, not shallow
This change addresses a concurrency problem, when two distinct threads
processing the same non-object template ended up modifying the same
HashResource object simultaneously, leading to TreeMap corruption, which
manifested itself as NullPointerException or compiler hang. The trigger
was a foreach() loop over a protected hash, which did not propagate
the protection down to the hash values; the stack trace leading for this
situation was:

     [panc] org.quattor.pan.dml.data.HashResource.put(HashResource.java:172)
     [panc] org.quattor.pan.dml.data.Resource.rput(Resource.java:131)
     [panc] org.quattor.pan.template.BuildContext.setLocalVariable(BuildContext.java:1298)
     [panc] org.quattor.pan.dml.operators.SetValue.execute(SetValue.java:170)
     [panc] org.quattor.pan.dml.operators.Assign.execute(Assign.java:65)
     [panc] org.quattor.pan.dml.operators.IfElse.execute(IfElse.java:109)
     [panc] org.quattor.pan.dml.DML.execute(DML.java:120)
     [panc] org.quattor.pan.dml.operators.Foreach.execute(Foreach.java:93)

Note there is a similar concurrency issue in
org.quattor.pan.type.RecordType.setDefaults(), which is _not_ addressed by this
patch. So this patch addresses, but does not completely solve, issues #6
and #56.

There is a visible side effect of this patch. There is the following
bit of code in the standard template library:

    variable DISK_VOLUME_PARAMS = {
      foreach (volume;params;DISK_VOLUME_PARAMS) {
        if ( (params['type'] == 'partition') &&
             (is_defined(DISK_PART_BY_DEV['changed_part_num'][params['device']])) ) {
          debug('Updating '+volume+' partition to new name/number: '+DISK_PART_BY_DEV['changed_part_num'][params['device']]);
          params['device'] = DISK_PART_BY_DEV['changed_part_num'][params['device']];
          params['final'] = true;
        };
      };
      SELF;
    };

The code iterates over and changes DISK_VOLUME_PARAMS, and expects the
changes to be visible in SELF. After this patch the above code will no
longer work, because DISK_VOLUME_PARAMS inside the DML block will have
copy-on-write semantics, changing it will not alter SELF. The foreach()
loop in the above code needs to be re-written to refrence SELF instead
of DISK_VOLUME_PARAMS.