panc: Deprecated should always return boolean true
This allows us to start removing the hundreds of instances of `... with {deprecated(...); true;}` in the schema.

It's not clear why it ever returned undef, but it appears to have been this way since the original rewrite in Java and looks like it was originally default boilerplate for new functions.
The documentation and comments claim it should be returning the string passed to it, but this has never been the case.